% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chart1.R
\name{create_chart1}
\alias{create_chart1}
\title{Create One Type Of Chart for Bulletin}
\usage{
create_chart1(
  datatable,
  plotfilename,
  plotwidth = 11,
  plotnrow = 2,
  plotheight = 5,
  plotxlabel = "x",
  plotylabel = "y",
  plottitle = "",
  plotsubtitle = "",
  plottag = "",
  wraplength = 35
)
}
\arguments{
\item{datatable}{placeholder}

\item{plotfilename}{placeholder}

\item{plotwidth}{placeholder}

\item{plotnrow}{placeholder}

\item{plotheight}{placeholder}

\item{plotxlabel}{placeholder}

\item{plotylabel}{placeholder}

\item{plottitle}{placeholder}

\item{plotsubtitle}{placeholder}

\item{plottag}{placeholder}

\item{wraplength}{placeholder}
}
\value{

}
\description{
Create One Type Of Chart for Bulletin
}
\examples{
# Not run:
# placeholder
}
