% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_charts.R
\name{create_charts}
\alias{create_charts}
\title{(Get Data From DHIS 2 And) Create Charts}
\usage{
create_charts()
}
\description{
(Get Data From DHIS 2 And) Create Charts
}
\examples{
# Not run
# create_charts()
}
