% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_bulletin.R
\name{draft_bulletin}
\alias{draft_bulletin}
\title{Create Draft DHIS 2 Federal Bulletin}
\usage{
draft_bulletin(
  title = "DHIS 2 federal bulletin",
  template = "federal",
  package = "dhis2bulletin"
)
}
\arguments{
\item{title}{File name of draft bulletin (default: "DHIS 2 federal bulletin")}

\item{template}{Template to use (default: "federal")}

\item{package}{Package that template is included in (default: "dhis2bulletin")}
}
\description{
Create Draft DHIS 2 Federal Bulletin
}
