% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_logindetails}
\alias{get_logindetails}
\title{Enter DHIS 2 Login Details}
\usage{
get_logindetails()
}
\description{
Enter DHIS 2 Login Details
}
\examples{
# Not run
# get_logindetails()
# Then use "get_pivottable" function
# Use username=admin password=district if using demo DHIS 2
}
