% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pivottable.R
\name{get_pivottable}
\alias{get_pivottable}
\title{Get Data from DHIS 2 Pivot Table (And Optionally Save It)}
\usage{
get_pivottable(url, params, basepath = NULL)
}
\arguments{
\item{url}{URL for DHIS 2 pivot table}

\item{params}{Parameters for username and password entered at knitting}

\item{basepath}{Name of data file to be saved}
}
\value{
A data.table
}
\description{
Get Data from DHIS 2 Pivot Table (And Optionally Save It)
}
