% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data1.R
\name{process_data1}
\alias{process_data1}
\title{Extract and Process Data from DHIS 2}
\usage{
process_data1(
  url,
  params,
  basepath = NULL,
  data_elements = NULL,
  wrap_length = 10
)
}
\arguments{
\item{url}{placeholder}

\item{params}{Parameters for username and password entered at knitting}

\item{basepath}{placeholder}

\item{data_elements}{placeholder}

\item{wrap_length}{placeholder}
}
\value{

}
\description{
Extract and Process Data from DHIS 2
}
