% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pivottable}
\alias{get_pivottable}
\title{Get Data from DHIS 2 Pivot Table (And Optionally Save It)}
\usage{
get_pivottable(url, filename = NULL)
}
\arguments{
\item{url}{URL for DHIS 2 pivot table}

\item{filename}{Name of data file to be saved}
}
\value{
A data.table
}
\description{
Get Data from DHIS 2 Pivot Table (And Optionally Save It)
}
\examples{
# Not run
# get_logindetails()
# Use username=admin password=district if using demo DHIS 2
# get_pivottable('https://play.dhis2.org/2.33.1/api/29/analytics.csv?dimension=pe:THIS_YEAR&dimension=lv8UXn17ZOm:weRMUzBs8T7&dimension=jp826jAJHUc:KUSvwZQsMSN;UmyRWILcoed;xNrDrDbJgnm;U0uJG4kydwE;kE8lP5t0b5R;yhg8oYU9ekY;oGktdmYkRNo;VCoSeRRVS1n;lLKpwhjd1dM;LzDaTmQYWcj;HDdnX6XqxIn;jWgEsdH87Jk;JZ3usxLEcc9;LqG1FnAUhyb;ZPtRFVLY40u;zmWJAEjfv59;Euvh58hLl61;RSFc8ADyKTw;SriP0jBXMr6;mcjC3qZgIkO;qk2KOBMX4Mf;HKU7L73im5r;TzwKbcw1nUK;lnLbEej0gwe;UAEhIWpoQFN;IUZ0GidX0jh;LgtuBcNaMB3;yHtsPZqpAxm;dUK38PhdUdV;HAraPb0v7ex;M2cth8EmrlT;GBHN1a1Jddh;KmwPVkjp7yl;oDkJh5Ddh7d&displayProperty=NAME&outputIdScheme=NAME')
}
